run-tests: download images only on x86-64
When running the tests on aarch64, the run-tests script tries to
download a pre-built image that is built for x86-64, and thus does not
run.

This change adds a check for the Docker daemon host platform and only
downloads the image if it will run.

Furthermore, the image dependency testbed/vim:24 is also built unless
the platform is x86_64, since it is also only provided for this
platform.

Test Plan:
Run `./run-tests` on both aarch64 and x86_64 without having
the required Docker images present (ie, `docker image ls` should not
list them).

Pull Request: https://github.com/apeschar/ale/pull/1